fix: missing-export errors
