+++ viewer/packages/api/index.ts
@@ -4,7 +4,7 @@
 export { Cognite3DViewer } from './src/public/migration/Cognite3DViewer';
 export { CogniteCadModel } from '@reveal/cad-model';
 
-export { ViewerState, ModelState } from './src/utilities/ViewStateHelper';
+export { ViewerState, ModelState, ClippingPlanesState } from './src/utilities/ViewStateHelper';
 
 export * from './src/public/types';
 export { CDF_TO_VIEWER_TRANSFORMATION } from './src/public/constants';
