+++ viewer/packages/tools/src/HtmlOverlay/HtmlOverlayTool.ts
@@ -76,7 +76,13 @@ export type HtmlOverlayToolClusteringOptions = {
   createClusterElementCallback: HtmlOverlayCreateClusterDelegate;
 };
 
+/**
+ * Options for the visualization of overlays
+ */
 export type HtmlOverlayToolOptions = {
+  /**
+   * Options for clustering the HTML overlays
+   */
   clusteringOptions?: HtmlOverlayToolClusteringOptions;
 };
 
@@ -93,7 +99,7 @@ type HtmlOverlayElement = {
 };
 
 /**
- 
+
  * Manages HTMLoverlays for {@link Cognite3DViewer}. Attaches HTML elements to a
  * 3D position and updates its position/visibility as user moves the camera. This is
  * useful to create HTML overlays to highlight information about key positions in the 3D model.
